Update sweetalert-dev.js
When the "sweetalert" is opening, if we click outside of the modal, the display will be canceled. Or if we click on "ok" before the end of the "animation", the modal will be closed but will not perform the action defined for this button.


The checking of the variable "modalIsVisible" was removed and this issue was resolved.